api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-multipart==0.0.6
httpx==0.25.1
PyJWT==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
scikit-learn==1.3.2
lightgbm==4.1.0
joblib==1.3.2
yfinance>=0.2.0
